pyproject.toml
[tool.poetry]
name = "erisbot"
version = "0.1.0"
description = "A fun Discord bot unlike others"
authors = ["No767 <73260931+No767@users.noreply.github.com>", "screencancer"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
python-dotenv = "^0.21.0"
aiodns = "^3.0.0"
Brotli = "^1.0.9"
orjson = "^3.8.1"
aiohttp = "^3.8.3"
"discord.py" = {extras = ["voice"], version = "^2.1.0"}
Pillow = "^9.3.0"
requests = "^2.28.1"
pytz = "^2022.6"
tortoise-orm = "^0.19.2"
aiosqlite = "^0.17.0"
numpy = "^1.23.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
